# SB-9-UserManagement/src/main/java/com/Entity/User.java
package com.Entity;

import java.time.LocalDate;
import java.util.Date;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Temporal;
import javax.persistence.TemporalType;

import lombok.Data;

@Entity
@Data
public class User {

	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	private int uid;

	private String uname;

	private String uemail;

	private String upassword;

	private String uaddress;

	private String gender;

	@Temporal(TemporalType.DATE)
	private Date date;

}