# SB-9-UserManagement/src/main/resources/AllAPIDoc.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:8005",
	"basePath": "/", 
...
	"paths": {
...
		"/exponent/adduser": {
			"post": {
				"tags": [
					"user-controller"
				],
				"summary": "AddUser",
				"operationId": "AddUserUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "user",
						"description": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/exponent/deleteUser": {
			"delete": {
				"tags": [
					"user-controller"
				],
				"summary": "deleteUserOnTheBasisOfID",
				"operationId": "deleteUserOnTheBasisOfIDUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "uid",
						"in": "query",
						"description": "uid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		},
		"/exponent/dummycall": {
			"post": {
				"tags": [
					"user-controller"
				],
				"summary": "DummyCall",
				"operationId": "DummyCallUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/exponent/getUsers": {
			"get": {
				"tags": [
					"user-controller"
				],
				"summary": "GetUserDetails",
				"operationId": "GetUserDetailsUsingGET",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/exponent/updateUser": {
			"put": {
				"tags": [
					"user-controller"
				],
				"summary": "updateRespectedUserUsingUID",
				"operationId": "updateRespectedUserUsingUIDUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "user",
						"description": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/User"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
...
		"User": {
			"type": "object",
			"properties": {
				"date": {
					"type": "string",
					"format": "date-time"
				},
				"gender": {
					"type": "string"
				},
				"uaddress": {
					"type": "string"
				},
				"uemail": {
					"type": "string"
				},
				"uid": {
					"type": "integer",
					"format": "int32"
				},
				"uname": {
					"type": "string"
				},
				"upassword": {
					"type": "string"
				}
			},
			"title": "User"
		},
...
	}
}